fix (uws): Update this to a fixed version and whole project to Node 8,10 (#47)
1) 'uws', due to some personal issues
(See: https://github.com/socketio/socket.io/issues/3319), the author
has destroyed it since 10.148.2. And after several times of trying. Only
10.148.0 can work very very properly in both Nodejs 8.x and 10.x. So use
this instead (Solution from
https://github.com/facebook/react-native/issues/17410).

2) Remove "please find another reliable library instead" in the
corresponding docs because we HAVE TO use 'uws', which is in the
engine.io library, it will check whether you are referring this lib or
not. If not, an error will be thrown out to you.
See:
https://github.com/socketio/engine.io/blob/6a16ea119280a02029618544d44eb515f7f2d076/lib/server.js#L107.

3) Upgrade the whole project to the latest version (Node 8, 10).